nxtweb/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
amqp==5.1.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
Brotli==1.0.9
bs4==0.0.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colosseum==0.2.0
contourpy==1.0.7
cron-descriptor==1.2.35
cryptography==38.0.1
cssselect==1.1.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.2
django-bootstrap-v5==1.0.11
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
django-localflavor==3.1
django-timezone-field==5.0
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.1
fake-useragent==1.1.3
fastapi==0.92.0
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gradio==3.13.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.3
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipdb==0.13.9
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
kombu==5.2.4
linkify-it-py==2.0.0
lxml==4.9.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==2.0.4
more-itertools==9.1.0
multidict==6.0.4
mutagen==1.46.0
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.4.12
numpy==1.24.0
openpyxl==3.0.10
orjson==3.8.7
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
paramiko==3.0.0
parse==1.19.0
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20220524
pdfplumber==0.7.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.31
proxy-tools==0.1.0
psutil==5.9.2
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pydantic==1.10.5
pydub==0.25.1
pyee==8.2.2
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.18.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-magic==0.4.27
python-multipart==0.0.6
python-stdnum==1.17
pytz==2022.4
pywebview==3.7.2
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.3.1
redis==4.5.3
regex==2022.10.31
requests==2.28.1
requests-html==0.10.0
rfc3986==1.5.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
starlette==0.25.0
terminado==0.16.0
tinycss2==1.1.1
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.26.1
typing_extensions==4.5.0
tzdata==2023.2
uc-micro-py==1.0.1
urllib3==1.26.12
uvicorn==0.20.0
vine==5.0.0
w3lib==2.0.1
Wand==0.6.11
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.3
widgetsnbextension==4.0.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
yt-dlp==2023.3.4
zipp==3.9.0
